Unblock tag indexing rule resources now that datadog-api-client-go#4184 merged
- Bump datadog-api-client-go to v2.60.1-0.20260608132748-ee5dce62079b, which
  includes the TagIndexingRule* types and MetricsApi methods generated from
  datadog-api-spec#5837.
- Fix six type mismatches between draft code and the actual generated types:
  int32→int64 for SetVersion/SetQueriedTagsWindowSeconds/SetQueriedWindowSeconds;
  *string nil check and deref for GetCreatedByHandleOk; addressable variable
  for GetRuleOrder pointer-receiver call.
- Remove WIP blocking comments from all three resource files.
- Add acceptance tests for datadog_tag_indexing_rule, _exemption, _order.
- Add examples for all three resources.
- Regenerate docs (also picks up llm_observability enum in dashboard/powerpack).

+++ datadog/fwprovider/resource_datadog_tag_indexing_rule.go
@@ -1,10 +1,5 @@
 package fwprovider
 
-// NOTE: This file references datadogV2.TagIndexingRule* types and MetricsApi methods
-// that do not yet exist in the current datadog-api-client-go version. This PR is a
-// draft, blocked on the datadog-api-client-go PR generated from datadog-api-spec#5837.
-// Once that client PR merges, update go.mod and remove this comment.
-
 import (
 	"context"
 
@@ -459,7 +454,7 @@ func (r *tagIndexingRuleResource) buildUpdateRequest(_ context.Context, state *t
 
 func buildOptionsFromModel(m *tagIndexingRuleOptionsModel) datadogV2.TagIndexingRuleOptions {
 	opts := datadogV2.NewTagIndexingRuleOptionsWithDefaults()
-	opts.SetVersion(int32(m.Version.ValueInt64()))
+	opts.SetVersion(m.Version.ValueInt64())
 
 	if m.Data != nil {
 		d := datadogV2.NewTagIndexingRuleOptionsDataWithDefaults()
@@ -469,7 +464,7 @@ func buildOptionsFromModel(m *tagIndexingRuleOptionsModel) datadogV2.TagIndexing
 		if m.Data.DynamicTags != nil {
 			dt := datadogV2.NewTagIndexingRuleDynamicTagsWithDefaults()
 			if !m.Data.DynamicTags.QueriedTagsWindowSeconds.IsNull() {
-				dt.SetQueriedTagsWindowSeconds(int32(m.Data.DynamicTags.QueriedTagsWindowSeconds.ValueInt64()))
+				dt.SetQueriedTagsWindowSeconds(m.Data.DynamicTags.QueriedTagsWindowSeconds.ValueInt64())
 			}
 			if !m.Data.DynamicTags.RelatedAssetTags.IsNull() {
 				dt.SetRelatedAssetTags(m.Data.DynamicTags.RelatedAssetTags.ValueBool())
@@ -489,7 +484,7 @@ func buildOptionsFromModel(m *tagIndexingRuleOptionsModel) datadogV2.TagIndexing
 				mm.SetNotUsedInAssets(m.Data.MetricMatch.NotUsedInAssets.ValueBool())
 			}
 			if !m.Data.MetricMatch.QueriedWindowSeconds.IsNull() {
-				mm.SetQueriedWindowSeconds(int32(m.Data.MetricMatch.QueriedWindowSeconds.ValueInt64()))
+				mm.SetQueriedWindowSeconds(m.Data.MetricMatch.QueriedWindowSeconds.ValueInt64())
 			}
 			if !m.Data.MetricMatch.UsedInAssets.IsNull() {
 				mm.SetUsedInAssets(m.Data.MetricMatch.UsedInAssets.ValueBool())
